fix(#8311): code fix

+++ src/panel/StockList.cpp
@@ -825,7 +825,7 @@ void StockList::onSize(wxSizeEvent& event)
     // calculate and apply diff:
     int diff = this->GetSize().GetWidth() - twidth;
     if (abs(diff) > 5) {
-        int diffdelta = diff / resizable_ids.size();
+        int diffdelta = diff / static_cast<int>(resizable_ids.size());
         for (colInfo col: resizable_ids) {
             this->SetColumnWidth(col.id, col.width + diffdelta);
         }
